kernel/KProve: converted into Guice injection. ProofDefinitionBuilder: moved Definition compiling logic here.
