+++ java-backend/src/main/java/org/kframework/backend/java/symbolic/InitializeRewriter.java
@@ -40,6 +40,7 @@ import org.kframework.utils.Stopwatch;
 import org.kframework.utils.errorsystem.KEMException;
 import org.kframework.utils.errorsystem.KExceptionManager;
 import org.kframework.utils.file.FileUtil;
+import org.kframework.utils.inject.RequestScoped;
 import org.kframework.utils.options.SMTOptions;
 import scala.Function1;
 import scala.Tuple2;
@@ -60,10 +61,10 @@ import static org.kframework.kore.KORE.*;
 /**
  * Created by dwightguth on 5/6/15.
  */
+@RequestScoped
 public class InitializeRewriter implements Function<org.kframework.definition.Definition, Rewriter> {
 
     private final FileSystem fs;
-    private final Stopwatch sw;
     private final GlobalOptions globalOptions;
     private final KExceptionManager kem;
     private final SMTOptions smtOptions;
@@ -91,11 +92,9 @@ public class InitializeRewriter implements Function<org.kframework.definition.De
             JavaExecutionOptions javaExecutionOptions,
             FileUtil files,
             InitializeDefinition initializeDefinition,
-            Stopwatch sw,
             KPrint kprint,
             Profiler2 profiler) {
         this.fs = fs;
-        this.sw = sw;
         this.globalOptions = globalOptions;
         this.kem = kem;
         this.smtOptions = smtOptions;
